refactor(baseball): 임의의 수 생성하는 기능 RandomNumber 클래스로 분리
